Move classes around
